Feature/refactor (#16)
* Update .gitignore

* Create detectors and trackers modules
