+++ .gitignore
@@ -131,3 +131,4 @@ dmypy.json
 # Project files/folders
 src/
 data/
+models/
